+++ src/lib/types/openapiDoc.ts
@@ -1,7 +1,14 @@
-import type { FromSchema } from 'json-schema-to-ts';
 import type { OASDocument } from 'oas/types';
 
-export const readmeAPIv2Oas = {
+/**
+ * This is a snapshot of the OpenAPI description for ReadMe APIv2.
+ *
+ * This is used both for typechecking as well as for runtime validation.
+ * We use  ajv to validate the user data against schemas in this document.
+ *
+ * @see {@link https://docs.readme.com/main/openapi/readme-api-v2-beta.json}
+ */
+const document = {
   openapi: '3.1.0',
   info: {
     description: 'Create beautiful product and API documentation with our developer friendly platform.',
@@ -5906,25 +5913,4 @@ export const readmeAPIv2Oas = {
   ],
 } as const satisfies OASDocument;
 
-export const categoryUriRegexPattern =
-  readmeAPIv2Oas.paths['/versions/{version}/guides'].post.requestBody.content['application/json'].schema.properties
-    .category.properties.uri.pattern;
-
-export const parentUriRegexPattern =
-  readmeAPIv2Oas.paths['/versions/{version}/guides'].post.requestBody.content['application/json'].schema.properties
-    .parent.properties.uri.pattern;
-
-type guidesRequestBodySchema =
-  (typeof readmeAPIv2Oas)['paths']['/versions/{version}/guides/{slug}']['patch']['requestBody']['content']['application/json']['schema'];
-
-/**
- * Derived from our API documentation, this is the schema for the `guides` object
- * as we send it to the ReadMe API.
- *
- * This is only for TypeScript type-checking purposes — we use ajv
- * to validate the user's schema during runtime.
- */
-export type GuidesRequestRepresentation = FromSchema<
-  guidesRequestBodySchema,
-  { keepDefaultedPropertiesOptional: true }
->;
+export default document;
